Moves templates dir to ui
