+++ lib/commands/init.js
@@ -118,8 +118,8 @@ Command.create({
 
     // Inject the template loading imports at the beginning of the template JS files
     _.each([
-      'imports/ui/templates/layouts/master-layouts/master-layout.js',
-      'imports/ui/templates/pages/default.js'
+      'imports/ui/layouts/master-layouts/master-layout.js',
+      'imports/ui/pages/default.js'
     ], function(jsPath) {
 
       var config = CurrentConfig.get();
