+++ lib/generators/route.js
@@ -3,7 +3,7 @@ var path = require('path');
 var RouteGenerator = Generator.create({
   name: 'route',
   aliases: ['r'],
-  usage: 'maka {generate, g}:{route, r} [path/]<name> [url] [--layout <layout>]',
+  usage: 'maka {generate, g}:{route, r} [path/]<name> [url]',
   description: 'Generate scaffolding for a Route.',
   examples: [
     'maka g:route todosIndex todos',
@@ -27,7 +27,7 @@ var RouteGenerator = Generator.create({
   name = this.fullCamelCase(name);
 
   var pathToTemplate = path.join(
-    '/imports/ui/templates/pages',
+    '/imports/ui/pages',
     opts.dir,
     this.fileCase(opts.resourceName),
     this.fileCase(opts.resourceName)
@@ -36,7 +36,7 @@ var RouteGenerator = Generator.create({
   var context = {
     name: name,
     url: url,
-    layout: opts.layout || "MasterLayout",
+    layout: "MasterLayout",
     templatePath: pathToTemplate,
     templateName: name
   };
