+++ lib/generators/template.js
@@ -18,13 +18,13 @@ Generator.create({
   name = name.replace(/\//, "-");
   name = this.fullCamelCase(name);
 
-  var projectPath = 'imports/ui/templates/pages';
+  var projectPath = 'imports/ui/pages';
 
   if(opts.component) {
-    projectPath = 'imports/ui/templates/components';
+    projectPath = 'imports/ui/components';
   }
   if(opts.layout) {
-    projectPath = 'imports/ui/templates/layouts';
+    projectPath = 'imports/ui/layouts';
   }
   if(opts.layout && opts.component) {
     this.logError("A template can not be a component and a layout at the same time.");
@@ -62,8 +62,8 @@ Generator.create({
   // Import the HTML and CSS
   this.injectAtBeginningOfFile(
     pathToTemplate + '.js',
-    "import './" + opts.resourceName + "." + config.engines.html + "';\n" +
-    "import './" + opts.resourceName + "." + config.engines.css + "';"
+    "import './" + this.fileCase(opts.resourceName) + "." + config.engines.html + "';\n" +
+    "import './" + this.fileCase(opts.resourceName) + "." + config.engines.css + "';"
   );
