+++ lib/templates/app/imports/startup/client/templates.js
@@ -1,5 +1,5 @@
 // This file is used to import all the templates into the project. 
 // Add new templates to the project by importing them here.
-import '/imports/ui/templates/pages/default.js';
-import '/imports/ui/templates/layouts/master-layouts/master-layout.js';
+import '/imports/ui/pages/default.js';
+import '/imports/ui/layouts/master-layouts/master-layout.js';
