+++ lib/templates/app/imports/ui/layouts/master-layouts/master-layout.html.jade
