+++ lib/templates/app/imports/ui/pages/default.html
@@ -1,0 +1,3 @@
+<template name="defaultPage">
+  <h1>Find me in app/imports/ui/pages/default.html</h1>
+</template>
