+++ lib/templates/app/imports/ui/pages/default.html.jade
@@ -1,0 +1,2 @@
+template(name="defaultPage")
+  h1 Find me in app/imports/ui/pages/default.jade
