--- lib/templates/app/imports/ui/templates/pages/default.html
@@ -1,3 +1,0 @@
-<template name="defaultPage">
-  <h2>Find me in app/imports/ui/templates/pages/default.html</h2>
-</template>
