--- lib/templates/app/imports/ui/templates/pages/default.html.jade
@@ -1,2 +1,0 @@
-template(name="defaultPage")
-  h2 Find me in app/imports/ui/templates/pages/default.jade
